Separate inlining-heuristic code into its own file. Implement a simple statement-counting inlining heuristic.
